App folder restructure
The structure is now broken out into pages, services, and components
