+++ app/models/Book.ts
