+++ app/models/Entry.ts
